Setup new flow for accurate fish simulation
Implement math for red dace population calculation per flow pulse
Create new flow system for fish
Update water flow style to average received values from all senders

+++ myWATERSHED/Assets/Scripts/CardSystem/CardAsset.cs
@@ -34,9 +34,9 @@ public class CardAsset : ScriptableObject
     public float m_asphaltDensity;
     public float m_erosionRate;
     public float m_landHeight;
-    public int m_pollutionLevel;
+    public float m_pollutionLevel;
     public float m_flowRate;
-    public int m_sewageLevel;
+    public float m_sewageLevel;
     public float m_sinuosity;
     public float m_shadeCoverage;
     public float m_turbidity;
